fix: inject scripts
