+++ apps/extension/src/inject/ethereum/eip6963.ts
@@ -1,4 +1,4 @@
-import { EthProvider } from "./types"
+import { type EthProvider } from "./types"
 
 interface EIP6963ProviderInfo {
   icon: string
