+++ apps/extension/src/ui/domains/Ramps/shared/RampsAccountPickerButton.tsx
@@ -1,8 +1,8 @@
 import { isAddressEqual } from "@talismn/crypto"
 import { ChevronRightIcon } from "@talismn/icons"
 import { TokenRatesList } from "@talismn/token-rates"
+import { HexString } from "@talismn/util"
 import { Account, getAccountGenesisHash, isAccountOwned } from "extension-core"
-import { HexString } from "extension-shared"
 import { FC, useCallback, useMemo } from "react"
 import { useTranslation } from "react-i18next"
 import { Drawer, useOpenClose } from "talisman-ui"
