+++ apps/extension/src/ui/domains/Settings/AddressBook/ContactCreateModal.tsx
@@ -1,8 +1,7 @@
 import { yupResolver } from "@hookform/resolvers/yup"
 import { isEthereumAddress } from "@polkadot/util-crypto"
 import { isAddressEqual, isSs58Address } from "@talismn/crypto"
-import { isValidSubstrateAddress } from "@talismn/util"
-import { HexString } from "extension-shared"
+import { HexString, isValidSubstrateAddress } from "@talismn/util"
 import { keyBy } from "lodash-es"
 import { useCallback, useEffect, useMemo } from "react"
 import { useForm } from "react-hook-form"
