+++ apps/extension/src/ui/domains/Settings/AddressBook/ContactEditModal.tsx
@@ -1,8 +1,8 @@
 import { yupResolver } from "@hookform/resolvers/yup"
 import { encodeAddressSs58 } from "@talismn/crypto"
 import { CopyIcon } from "@talismn/icons"
+import { HexString } from "@talismn/util"
 import { isAccountAddressSs58, isAddressCompatibleWithNetwork } from "extension-core"
-import { HexString } from "extension-shared"
 import { keyBy } from "lodash-es"
 import { FC, useCallback, useMemo } from "react"
 import { useForm } from "react-hook-form"
