+++ apps/extension/src/ui/util/getFrontendTypeRegistry.ts
@@ -4,7 +4,7 @@ import { getSpecAlias, getSpecTypes } from "@polkadot/types-known/util"
 import { hexToNumber } from "@polkadot/util"
 import { HexString } from "@polkadot/util/types"
 import { DotNetwork } from "@talismn/chaindata-provider"
-import { getMetadataFromDef, getMetadataRpcFromDef } from "extension-shared"
+import { getMetadataFromDef, getMetadataRpcFromDef } from "extension-core"
 
 import { api } from "@ui/api"
