+++ packages/crypto/src/address/index.ts
@@ -2,3 +2,4 @@ export * from "./encoding"
 export * from "./addressFromPublicKey"
 export * from "./normalizeAddress"
 export * from "./isAddressEqual"
+export * from "./isValidAddress"
