+++ packages/crypto/src/address/isValidAddress.ts
@@ -1,0 +1,10 @@
+import { normalizeAddress } from "./normalizeAddress"
+
+export const isValidAddress = (address: string): boolean => {
+  try {
+    normalizeAddress(address)
+    return true
+  } catch {
+    return false
+  }
+}
