+++ packages/extension-core/src/index.ts
@@ -68,3 +68,4 @@ export { privacyRoundCurrency } from "./util/privacyRoundCurrency"
 export * from "./domains/nfts/exports"
 export * from "./domains/keyring/exports"
 export * from "./domains/transactions/exports"
+export * from "./domains/metadata/helpers"
