+++ packages/extension-core/src/libs/QrGenerator/index.ts
@@ -1,11 +1,12 @@
 import { Keyring } from "@polkadot/keyring"
 import { assert, hexToU8a, u8aConcat, u8aToU8a } from "@polkadot/util"
 import { DotNetwork } from "@talismn/chaindata-provider"
-import { getMetadataRpcFromDef, log } from "extension-shared"
+import { log } from "extension-shared"
 
 import { appStore } from "../../domains/app/store.app"
 import { passwordStore } from "../../domains/app/store.password"
 import { keyringStore } from "../../domains/keyring/store"
+import { getMetadataRpcFromDef } from "../../domains/metadata/helpers"
 import { SignerPayloadGenesisHash } from "../../domains/signing/types"
 import { chainConnector } from "../../rpcs/chain-connector"
 import { chaindataProvider } from "../../rpcs/chaindata"
