+++ packages/extension-core/src/util/getMetadataDef.ts
@@ -2,10 +2,11 @@ import { assert, isHex } from "@polkadot/util"
 import { HexString } from "@polkadot/util/types"
 import { DotNetwork, DotNetworkId, NetworkId } from "@talismn/chaindata-provider"
 import { fetchBestMetadata } from "@talismn/sapi"
-import { DEBUG, encodeMetadataRpc, log } from "extension-shared"
+import { DEBUG, log } from "extension-shared"
 
 import { sentry } from "../config/sentry"
 import { db } from "../db"
+import { encodeMetadataRpc } from "../domains/metadata/helpers"
 import { metadataUpdatesStore } from "../domains/metadata/metadataUpdates"
 import { TalismanMetadataDef } from "../domains/substrate/types"
 import { chainConnector } from "../rpcs/chain-connector"
