+++ packages/extension-core/src/util/getTypeRegistry.ts
@@ -2,8 +2,9 @@ import { typesBundle } from "@polkadot/apps-config/api"
 import { Metadata, TypeRegistry } from "@polkadot/types"
 import { getSpecAlias, getSpecTypes } from "@polkadot/types-known/util"
 import { hexToNumber, isHex } from "@polkadot/util"
-import { getMetadataFromDef, getMetadataRpcFromDef, log } from "extension-shared"
+import { log } from "extension-shared"
 
+import { getMetadataFromDef, getMetadataRpcFromDef } from "../domains/metadata/helpers"
 import { chaindataProvider } from "../rpcs/chaindata"
 import { getMetadataDef } from "./getMetadataDef"
