--- packages/extension-shared/src/domains/index.ts
@@ -1,1 +1,0 @@
-export * from "./substrate"
