--- packages/extension-shared/src/domains/substrate/index.ts
@@ -1,1 +1,0 @@
-export * from "./types"
