--- packages/extension-shared/src/domains/substrate/types.ts
@@ -1,5 +1,0 @@
-import type { MetadataDef as PjsMetadataDef } from "@polkadot/extension-inject/types"
-
-export interface TalismanMetadataDef extends PjsMetadataDef {
-  metadataRpc?: `0x${string}`
-}
