+++ packages/extension-shared/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./constants"
-export * from "./domains"
 export * from "./log"
 export * from "./util"
