--- packages/extension-shared/src/util/decodeMetaCalls.ts
@@ -1,4 +1,0 @@
-import { base64Decode } from "@polkadot/util-crypto"
-
-// these are stored as base64 for lower storage size
-export const decodeMetaCalls = (encoded: string) => base64Decode(encoded)
