+++ packages/extension-shared/src/util/index.ts
@@ -1,11 +1,2 @@
-export * from "./decodeMetaCalls"
-export * from "./decodeMetadataRpc"
-export * from "./encodeMetadataRpc"
-export * from "./getAddressType"
-export * from "./getChainAddressType"
-export * from "./getMetadataFromDef"
-export * from "./getMetadataRpcFromDef"
 export * from "./isTalismanHostname"
 export * from "./isTalismanUrl"
-export * from "./isValidAddress"
-export * from "./isHexString"
