+++ packages/util/src/index.ts
@@ -31,3 +31,4 @@ export * from "./isTruthy"
 export * from "./isAbortError"
 export * from "./getSharedObservable"
 export * from "./keepAlive"
+export * from "./isHexString"
