+++ packages/util/src/isHexString.ts
